backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.1.1
google-cloud-firestore==2.21.0
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.28.0
google-genai==1.31.0
pytest==8.4.1
